Added the principal gems used and some basic configuration.
ADD: Gems to Gemfile
Added some gems that are going to be used, like Devise, MongoID or
Bootstrap. Also, cleanes some unused gems from the original Gemfile.

ADD: Configuration files
Added the configuration files for MongoID and Devise.

MODIFY: Application configuration files
Modified to delete all the references to ActiveRecord, because it's
no longer used in favor of MongoID as database driver.

MODIFY: Default JS and CSS file.
Deleted the reference to turbolinks libraries and the initial
comments in the files.