Modified RDoc style and configuration
MODIFY: RDoc style and configuration
The RDoc gem, as long with the theme gem hanna-bootstrap were added
to the Gemfile in the :doc group.
Also, the RDoc default parameters to avoid parsing non-ruby files,
configure the template, etc.

DELETE: Some unused stuff
README.rdoc was unused in favor of README.md, and config/database.yml
was used only with ActiveRecord, so it's no longer needed because
the project uses MongoID.