Added Devise model & views
MODIFY: Moved UI strings to localization file
Added the UI strings in the localization file for English language.

ADD: Devise model & views
Runned the generator of User model to use it with Devise to provide
user authentication and management.
Runned the view generator to being able to modify the views to match
app's style.