Added more complex SMTP configuration and a fix for Devise.
ADD: More complex SMTP configuration
Allowing to set if the connection with the SMTP server is using
a plain, SSL/TLS or STARTTLS security mode.

MODIFY: Comments in default configuration file
Added more detailed comments to provide the available values
in every configuration flag.

FIX: Bug with Devise and Rails 4.1
There is a bug in session cookie encoding that causes a fail
retrieving the user session from MongoDB.
It was  solved using the indications readed here:
https://github.com/plataformatec/devise/issues/2949#issuecomment-58162176