Added default layout for e-mails.
ADD: Layout for e-mails
Added a layout in HTML and text version for the e-mails generated
by the app.

MODIFY: Setted configuration variables for assets host.
Because there is a possibility of using a external CDN for the
assets, there is a new configuration variable in config/watchr.yml
file for setting the assets host used in both web and e-mail urls
for assets.