Added installation process used when there isn't an administrator
ADD: Installation process
Applyed filters to determine when the app need to run the installation
process to create an administrator account.
The process shows a form with required data to create an account.
The password is generated by the system and e-mailed to the user.

ADD: Password generation functionality to User model
Added a function to generate random secure passwords, used at the
moment in the installation process.