Modified devise functionality to work better
ADD: Notification when the token for password change fails
Now, when an user tries to change his password with an outdated
token, a warning is shown with a button to request a new token.

MODIFY: Added error indicators and requirements to devise views
Added error indicators using devise, indications and Bootstrap classes.
Also, the required fields has been marked to be required by the
browser before the send.

ADD: Title to every page
Added the title of every section to be shown in layouts.

ADD: Devise particular JavaScript
To avoid loading too much unnecesary code when the application
main JS file starts to grown.