Added more descripting error messages
ADD: Detailed error messages in Devise and Installation views.
Now, it shows the requeriments of a password, name and e-mail, so
if something that the user enters doesn't fir the requeriments, the
user can know why the operation fails.

MODIFIED: Name validation rules
The name length now must be between 2 and 30 characters.