Added language preference in User model
ADD: Language preference for User model.
Added a field for language, validations that these language exists
in localization files, and a before action function to be run
before every action in the app that selects the locale to use based
on user preferences (if the user isn'y logged in, use the default
locale of the system.