Added the possibility to change a user password.
ADD: User password change.
Now an administrator can choose to generate a new password and
the system will auto-generate and send it to the user (the administrator
can't see any password).

MODIFY: before_action filters to check privileges.
Now the check of privileges (if it's the correct or upper) is made
in the user model. Also, added some messages in the redirections
when the user isn't allowed to access a feature.

MODIFY: Privilege requeriments to do user actions
Normal users can only search and see the details of other users.
Administrator users can create, update and destroy other users.