Added Preferences view, and changed the sidebars
ADD: Preferences view
Using multiple panes and forms. Also added translations to all the
texts.

MODIFY: Users' sidebar
Added an external box, and changed the column sizes

MODIFY: Some visual bugs
Fixed problems when the users list is showed in a small device.
Also, added a class='active' to the Users element in the navbar
when the UsersController is being used.