Finished the preferences saving and added somo HD images
MODIFY: Save process for preferences
Added the functionality to save preferences, checking the current
password and the data entered.
If an error occurred, the modify form is shown again with information
of the error.

ADD: Some HD images
Added the logo images in Retina format for Apple devices (2x resolution).