Added some tests for models, mailers and controllers
ADD: User model tests
They test all the validations and functions of the User model,
welcome e-mail sending and filters application.

ADD: UserMailer tests
They tests the from, to and subject fields of the e-mails delivered
by the mailer.

ADD: InstallationController tests
They tests the limitaion of the installation, the data check performed
after the administrator creation, and the behaviour of the installation
process.

ADD: PanelController test
They only test (by now) the redirection in case of an unauthorized user.