Added some protections against auto-deleting administrators.
MODIFY: UsersControllers actions and views
Now, if an administrator user tries to destroy it's own account,
the system doesn't allow him and shows an error message. Also,
the delete buttons in his own account are disabled (in index and
show views).

Users can't change now it's privilege level. If an administrator
is going to be a less-privileged user, another administrartor
must downgrade hiss privilege level.