Added new and create actions of HostsController
ADD: New and Create actions of HostsController
Added actions to create new hosts on the application. Also added
the hosts form, and Strong Parameters on HostsController to avoid
mass assignment on Host model.

MODIFY: Navbar distribution
Now there is a new entry, Configuration. there is the management of
Hosts, Services and Alerts.