Added some translations and actions of HostsController
ADD: Translations of notices and errors
Added some translations of notices and errors shown by the controller
of hosts.

MODIFY: Error handling in hosts form
Modified to get the difference between an address invalid error and
an address repeated error.

ADD: Actions for edit, update and destroy in HostsController
Added actions that allow to show an edit form, save the updated data
and destroy a host from the database.