Added Probe architecture and a dummy probe
ADD: Probe architecture
The probes are stored insed /lib/probes folder. There is a folder for each
probe, named with the identificator of the probe (the "name").
Inside this folder there is a .rb file with this identificator. In this
file will be the functions that allows to do the probe, check configuration,
intervals, etc.
In this folder are two more folders: locales and views, used for localizations
of the texts used in the probe and the views needed for configuration.

It's important to follow the convention, because the main module of probes,
called Watchr::Probes, autoload them on boot time.

Also, there is a main class called Watchr::Probe, loaded with 'probes/probe',
that include a set of helper functions to help a probe to communicate with
the Watchr application.

ADD: Dummy probe
Added a dummy probe for testing pourposes.

MODIFY: Boot process to include probe autoloading
Added the autoloading of probe classes and the locales of each probe.