Added Service model and some functions in Probes class.
ADD: Service model
Added fields and validations to store the expected data of a service,
as long as integration with the Probes system developed.

MODIFY: New functions in Probes class
Functions to get the registered probes identificators, check if a
probe identificator is valid, or get a class object of a probe.