Added complete Service model and modified the host types
ADD: Complete Service model
Finished the Service model with the validations and functions needed
to check resume types, priorites, probes, etc.

MODIFY: Host types
Modified the host types in Host model to became symbols, in orden
to be more readable. This isn't extended to other parameters in
other models where the cardinality of the numbers means an specific
order (like in services priorities, when the lowest number indicates
the highest priority).