Added probe configuration interface.
ADD: Probe configuration interface.
The probe configuration interfaces are generated by the probes themselves
and are isnerted in the creation/edition form, and refreshed when
the user changes the probe selection field.

Also, methods to check if the configuration is valid has been implemented
and the system is now able to check the probe's config and warn the
user if something is wrong

ADD: Protection against uninstalled probes.
If a service has assigned an uninstalled probe, an error is shown
in the index table and in the show view of the service.