Added checks on host assignation to service
ADD: Checks on host-service management
Added a filter to control the existence of hosts before the assign
of hosts to service. When index_hosts, new_host and delete_host actions
are called, if no host exists the user is redirected back and an
error message is shown.

Also, added privilege level checks to the new actions.

MODIFY: Host and service show views
Added a list of services and hosts assigned to both of them.

MODIFY: Configuration sidebar and navbar link.
Now the services management is more important than hosts management.