Added the User-Alert relationship.
ADD: User-Alert relation.
Added a relation between User and Alert model to save the users that
must be notified if an alert gets activated or deactivated.

ADD: User-Alert relation management interface
Added the actions needed to manage this relation (specify the users to
be notified when the status of an alert changes) using the web interface.