Added async send for UserMailer e-mails and i18n for DeviseAsync
ADD: Async send for UserMailer e-mails.
Added an UserMailerWorker to do the tasks required in order to send
the user related e-mails, taking care of the language specified by
the user.

This was done with a worker to avoid duplicated code for the i18n part
in both methods of the mailer.

MODIFY: i18n support for Devise::Mailer
Added in the initialization file of Devise Async because it's a problem
derivated from the use of this gem.

Now the mailer takes care of the language defined by the user in order
to send a correct e-mail.

MODIFY: User and UsersController tests
To test that the background jobs are sended in replacement of the
e-mail send test that was done before.