Added Result and HostResult models
ADD: Result model
Used to store all the results of a probe in all the hosts of the service
to monitor. Ir uses HostResult documents as embedded to store particular
host results of the probe.

It includes functions to get the specific result of a host, or to know
if there is an error, getting the global value of the probes to the hosts,
etc.

ADD: HostResult model
Used to store a particular probe result done to a host. It includes the
status (ok or error), an error message (of the status is error) and a
float value (if the status is ok).

Also, there are checks to verify that the fields are correct and there
isn't an error message if the status isn't error.