Bugfixes and inactive host check
MODIFY: Bugfix in services form.
There is an incorrect translation in clean_interval field.

MODIFY: Check inactive hosts in ServiceProbeWorker
Now it only performs the probe over active hosts. If every host assigned
is inactive, no reuslt is saved.

MODIFY: Bugfix in ServiceProbeWorker
When the probe of a service becomes unavailable, the service gets becomes
inactivated and the job stops until the problem gets solved or the
application restarted.