Added some probes.
MODIFY: Inclusion of Probe#description_html method.
Used to return an HTML valid (and safe) description of a probe.

ADD: ICMP ping probe
It's a probe to test the connectivity of a remote host using the ICMP
protocol.
Due to the requeriment of open a RAW socket (no port is used), it needs
root privileges to work, so Sidekiq must be run in root mode.

ADD: External ping probe
It's a probe similar to the ICMP ping probe, but it uses the ping application
provided by the operating system, so it doesn't need root privileges
to work.
It also allows to define how many pings to send and the interval between them.

MODIFY: Gemfile
Added net-ping gem to support the ping functionality