Added UDP ping and TCP connection probes
ADD: UDP ping
Added a probe to test UDP connectivity using a ping test. It sends
a message to a UDP port and waits for the same message in response.

ADD: TCP connection probe
Added a probe to test a TCP port connectivity. It tries to connect
to a determined TCP port of the remote host.
If the probe is ok, the time to connect is saved as value. If not,
an error message is saved.