Added HTTP petition probe.
ADD: HTTP petition probe
It allows to perform an HTTP Get petition to a remote host and
save the response time.
It allows to define the remote port (80 by default), the path of the
URL, if the user-agent has to follow redirections, and if the
petition must be a GET or a HEAD petition.