Added service status and overall results in the monitoring section.
ADD: Service status.
Used to know the status of a service. It is OK if all the hosts
assigned have returned a value in the last probe execution.

If some hosts failed, it returns a WARNING status, and if all of
them failed, it returns an ERROR status.

ADD: Overall results section
Thios section of the Monitoring module shows a brief resume of
every service active and with assigned hosts.

ADD: Fluid multiple-row classes
Used in the overall section, it provides a fluid grid that allows
to get a pinterest-like distribution without vertical spaces between
rows due to diferente heigths.