Changes to service results view
ADD: Tables with service information and results
Added two tables in service results view: one with information
about the service configuration and one with the lastest
results measured in each host.

ADD: Live reload of results
The table of results is now auto-reloaded using the service
interval to maintain the data updated.

MODIFY: Added a tfoot class in tables
Used to show results (like the global value measured in the last
results).