Added units of the results to the probes and general changes
ADD: Units to probes
Added unit information in th probes to being able to interpret
the results in a good way.
The default unit is th second, but can be changed as can be seen
in Dummy probe.

ADD: Initial host results page
To be continued.

MODIFY: Improved charts
Changes to make the charts more extensive (especifying the div in
which they are going to be) and more i18n-friendly.

MODIFY: general UI changes
Added units in results fields and reducted the number of translations
needed.