Added Inactive result status
ADD: Inactive result status
Used to save information of a inactive result when trying to execute
the probes over an inactive host. It's HostResult is saved with
Inactive status and it can be shown in the results pages (and showed in
the charts as a 0.0 value).
Also, inactive results don't count as errors, so a service with a lot
of all hosts inactived isn't in ERROR state.

ADD: Sidekiq link and protection
Added sidekiq access link in user's dropdown, only for administrators.
Also, the action has been protected.

MODIFY: Helper functions to generate status
Modified to support the status of a service and a HostResult.
Also, changed the results' panels views to use the new functions.