Added tests and small changes
ADD: Result tests
Testing all the functionality of the model and the validations.

MODIFY: Some validations for Result model
Added validation to check if there is more than one HostResult of
the same host, and to check if there is any HostResult of a host
not linked to the service.

MODIFY: Chart colours
Modified to avoid using non-flat designs