Added alert activation/deactivation and host remove compatibility
ADD: Alert activation/deactivation
There is now a method in Alert model that is called when a resul object
os saved. in that function the results are checked against the
properties of the Alert to open/maintain/close an AlertRecord.

MODIFY: Host removal function
When a host is removed from a Service, it is removed also from the
Alerts that are linked to that Service. And when a Host is removed from
the Alert, all their AlertRecords are closed automatically.