Using HighStock, improves in JSON generation and in UX
MODIFY: Use of HighStock
Changed from HighCharts to HighStock in order to avoid problems with
large number of points.

MODIFY: Improve JSON generation
Removed latest changes to improve JSON generation

MODIFY: Number of points in charts
Changed some charts to make a faster loading.

MODIFY: Limit of points to return in JSON generation
Parameter :limit in URL of JSON API limits the number of points to
return in the JSON, so it can be created quickly.